+++ src/GRA.Controllers/ViewModel/Home/DashboardViewModel.cs
@@ -41,7 +41,8 @@ namespace GRA.Controllers.ViewModel.Home
         public int? PercentComplete { get; set; }
         public string ProgressMessage { get; set; }
 
-        public IEnumerable<Badge> Badges { get; set; }
+        public ICollection<UserLog> UserLogs { get; set; }
+
         public ICollection<AvatarElement> AvatarElements { get; set; }
 
         public string ProgramName { get; set; }
@@ -51,7 +52,8 @@ namespace GRA.Controllers.ViewModel.Home
         public Carousel Carousel { get; set; }
         public SiteStage SiteStage { get; set; }
 
-        public bool HasPendingDonationQuestion { get; set; }
+        public bool HasPendingVendorCodeQuestion { get; set; }
         public DateTime? VendorCodeExpiration { get; set; }
+        public ICollection<Event> UpcomingStreams { get; set; }
     }
 }