+++ src/GRA.Data.SqlServer/Migrations/SqlServerContextModelSnapshot.cs
@@ -830,6 +830,8 @@ namespace GRA.Data.SqlServer.Migrations
                     b.Property<string>("Email")
                         .IsRequired();
 
+                    b.Property<DateTime?>("SentAt");
+
                     b.Property<string>("SignUpSource")
                         .IsRequired();
 
@@ -879,19 +881,23 @@ namespace GRA.Data.SqlServer.Migrations
                     b.Property<int>("CreatedBy");
 
                     b.Property<string>("Description")
+                        .IsRequired()
                         .HasMaxLength(255);
 
                     b.Property<int>("EmailsSent");
 
                     b.Property<string>("FromAddress")
+                        .IsRequired()
                         .HasMaxLength(255);
 
                     b.Property<string>("FromName")
+                        .IsRequired()
                         .HasMaxLength(255);
 
                     b.Property<int>("SiteId");
 
                     b.Property<string>("Subject")
+                        .IsRequired()
                         .HasMaxLength(255);
 
                     b.HasKey("Id");
@@ -951,6 +957,10 @@ namespace GRA.Data.SqlServer.Migrations
 
                     b.Property<bool>("IsCommunityExperience");
 
+                    b.Property<bool>("IsStreaming");
+
+                    b.Property<bool>("IsStreamingEmbed");
+
                     b.Property<bool>("IsValid");
 
                     b.Property<string>("Name")
@@ -970,6 +980,11 @@ namespace GRA.Data.SqlServer.Migrations
                     b.Property<int>("SiteId");
 
                     b.Property<DateTime>("StartDate");
+
+                    b.Property<DateTime?>("StreamingAccessEnds");
+
+                    b.Property<string>("StreamingLinkData")
+                        .HasMaxLength(255);
 
                     b.HasKey("Id");
 
@@ -1418,6 +1433,8 @@ namespace GRA.Data.SqlServer.Migrations
 
                     b.Property<int>("UserId");
 
+                    b.Property<int?>("VendorCodeId");
+
                     b.HasKey("Id");
 
                     b.HasIndex("TriggerId");
@@ -1554,6 +1571,8 @@ namespace GRA.Data.SqlServer.Migrations
                     b.Property<DateTime>("CreatedAt");
 
                     b.Property<int>("CreatedBy");
+
+                    b.Property<bool>("IsDeleted");
 
                     b.Property<int?>("KitId");
 
@@ -2716,6 +2735,23 @@ namespace GRA.Data.SqlServer.Migrations
                     b.ToTable("UserFavoriteChallenges");
                 });
 
+            modelBuilder.Entity("GRA.Data.Model.UserFavoriteEvent", b =>
+                {
+                    b.Property<int>("UserId");
+
+                    b.Property<int>("EventId");
+
+                    b.Property<DateTime>("CreatedAt");
+
+                    b.Property<int>("CreatedBy");
+
+                    b.HasKey("UserId", "EventId");
+
+                    b.HasIndex("EventId");
+
+                    b.ToTable("UserFavoriteEvents");
+                });
+
             modelBuilder.Entity("GRA.Data.Model.UserLog", b =>
                 {
                     b.Property<int>("Id")
@@ -2740,6 +2776,8 @@ namespace GRA.Data.SqlServer.Migrations
 
                     b.Property<string>("Description");
 
+                    b.Property<int?>("EventId");
+
                     b.Property<bool?>("HasBeenViewed");
 
                     b.Property<bool>("IsDeleted");
@@ -2747,6 +2785,8 @@ namespace GRA.Data.SqlServer.Migrations
                     b.Property<int?>("PointTranslationId");
 
                     b.Property<int>("PointsEarned");
+
+                    b.Property<int?>("TriggerId");
 
                     b.Property<int>("UserId");
 
@@ -2814,8 +2854,11 @@ namespace GRA.Data.SqlServer.Migrations
                         .ValueGeneratedOnAdd()
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
+                    b.Property<int?>("BranchId");
+
                     b.Property<string>("Code")
-                        .IsRequired();
+                        .IsRequired()
+                        .HasMaxLength(255);
 
                     b.Property<DateTime>("CreatedAt");
 
@@ -2823,7 +2866,19 @@ namespace GRA.Data.SqlServer.Migrations
 
                     b.Property<DateTime>("DateUsed");
 
+                    b.Property<string>("Details")
+                        .HasMaxLength(255);
+
+                    b.Property<string>("EmailAwardAddress")
+                        .HasMaxLength(255);
+
+                    b.Property<DateTime?>("EmailAwardReported");
+
+                    b.Property<DateTime?>("EmailAwardSent");
+
                     b.Property<bool?>("IsDonated");
+
+                    b.Property<bool?>("IsEmailAward");
 
                     b.Property<bool>("IsUsed");
 
@@ -2851,6 +2906,8 @@ namespace GRA.Data.SqlServer.Migrations
                         .ValueGeneratedOnAdd()
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
+                    b.Property<bool>("AwardPrizeOnShipDate");
+
                     b.Property<DateTime>("CreatedAt");
 
                     b.Property<int>("CreatedBy");
@@ -2865,12 +2922,16 @@ namespace GRA.Data.SqlServer.Migrations
                     b.Property<string>("DonationMessage")
                         .HasMaxLength(255);
 
-                    b.Property<string>("DonationOptionMail")
+                    b.Property<string>("DonationSubject")
+                        .HasMaxLength(255);
+
+                    b.Property<string>("EmailAwardMail")
                         .HasMaxLength(1250);
 
-                    b.Property<string>("DonationOptionSubject");
+                    b.Property<string>("EmailAwardMessage")
+                        .HasMaxLength(255);
 
-                    b.Property<string>("DonationSubject")
+                    b.Property<string>("EmailAwardSubject")
                         .HasMaxLength(255);
 
                     b.Property<DateTime?>("ExpirationDate");
@@ -2883,6 +2944,11 @@ namespace GRA.Data.SqlServer.Migrations
                         .IsRequired()
                         .HasMaxLength(255);
 
+                    b.Property<string>("OptionMail")
+                        .HasMaxLength(1250);
+
+                    b.Property<string>("OptionSubject");
+
                     b.Property<int>("SiteId");
 
                     b.Property<string>("Url")
@@ -2893,6 +2959,22 @@ namespace GRA.Data.SqlServer.Migrations
                     b.HasIndex("SiteId");
 
                     b.ToTable("VendorCodeTypes");
+                });
+
+            modelBuilder.Entity("GRA.Data.Model.VendorCodeTypeText", b =>
+                {
+                    b.Property<int>("LanguageId");
+
+                    b.Property<int>("VendorCodeTypeId");
+
+                    b.Property<string>("EmailAwardInstructions")
+                        .HasMaxLength(1000);
+
+                    b.HasKey("LanguageId", "VendorCodeTypeId");
+
+                    b.HasIndex("VendorCodeTypeId");
+
+                    b.ToTable("VendorCodeTypeTexts");
                 });
 
             modelBuilder.Entity("Microsoft.AspNetCore.DataProtection.EntityFrameworkCore.DataProtectionKey", b =>
@@ -3531,6 +3613,19 @@ namespace GRA.Data.SqlServer.Migrations
                         .OnDelete(DeleteBehavior.Restrict);
                 });
 
+            modelBuilder.Entity("GRA.Data.Model.UserFavoriteEvent", b =>
+                {
+                    b.HasOne("GRA.Data.Model.Event", "Event")
+                        .WithMany()
+                        .HasForeignKey("EventId")
+                        .OnDelete(DeleteBehavior.Restrict);
+
+                    b.HasOne("GRA.Data.Model.User", "User")
+                        .WithMany()
+                        .HasForeignKey("UserId")
+                        .OnDelete(DeleteBehavior.Restrict);
+                });
+
             modelBuilder.Entity("GRA.Data.Model.UserLog", b =>
                 {
                     b.HasOne("GRA.Data.Model.User", "User")
@@ -3591,6 +3686,19 @@ namespace GRA.Data.SqlServer.Migrations
                     b.HasOne("GRA.Data.Model.Site", "Site")
                         .WithMany()
                         .HasForeignKey("SiteId")
+                        .OnDelete(DeleteBehavior.Restrict);
+                });
+
+            modelBuilder.Entity("GRA.Data.Model.VendorCodeTypeText", b =>
+                {
+                    b.HasOne("GRA.Data.Model.Language", "Language")
+                        .WithMany()
+                        .HasForeignKey("LanguageId")
+                        .OnDelete(DeleteBehavior.Restrict);
+
+                    b.HasOne("GRA.Data.Model.VendorCodeType", "VendorCodeType")
+                        .WithMany()
+                        .HasForeignKey("VendorCodeTypeId")
                         .OnDelete(DeleteBehavior.Restrict);
                 });
 #pragma warning restore 612, 618