+++ src/GRA.Data/Model/VendorCode.cs
@@ -11,6 +11,7 @@ namespace GRA.Data.Model
         public int VendorCodeTypeId { get; set; }
         public VendorCodeType VendorCodeType { get; set; }
         [Required]
+        [MaxLength(255)]
         public string Code { get; set; }
         [Required]
         public bool IsUsed { get; set; }
@@ -20,5 +21,15 @@ namespace GRA.Data.Model
         public DateTime? OrderDate { get; set; }
         public DateTime? ShipDate { get; set; }
         public bool? IsDonated { get; set; }
+        public bool? IsEmailAward { get; set; }
+
+        [MaxLength(255)]
+        public string EmailAwardAddress { get; set; }
+
+        public DateTime? EmailAwardReported { get; set; }
+        public DateTime? EmailAwardSent { get; set; }
+        [MaxLength(255)]
+        public string Details { get; set; }
+        public int? BranchId { get; set; }
     }
 }