+++ src/GRA.Domain.Model/Filters/ChallengeFilter.cs
@@ -5,6 +5,7 @@
         public bool? Favorites { get; set; }
         public int? FavoritesUserId { get; set; }
         public int? GroupId { get; set; }
+        public bool? IsCompleted { get; set; }
 
         public ChallengeFilter(int? page = null, int? take = null) : base(page) { }
     }