+++ src/GRA/SiteSettingDefinitions.cs
@@ -48,6 +48,14 @@ namespace GRA
                     Category = typeof(SiteSettingKey.Events).Name,
                     Format = SiteSettingFormat.Boolean
                 },
+                [SiteSettingKey.Events.HideUntilProgramOpen] =
+                new SiteSettingDefinition
+                {
+                    Name = "Hide until program opens",
+                    Info = "Put any text here to hide events until the program is open",
+                    Category = typeof(SiteSettingKey.Events).Name,
+                    Format = SiteSettingFormat.Boolean
+                },
                 [SiteSettingKey.Events.RequireBadge] =
                 new SiteSettingDefinition
                 {
@@ -150,6 +158,22 @@ namespace GRA
                     Name = "Check for site.css and site.js changes on disk",
                     Info = "How often (in minutes) To check if site.css and site.js have changed on disk. Set to 0 when working on the files so that they will reload on each page load, set to a higher number when you are editing them infrequently. An empty setting will default to 60 minutes",
                     Category = typeof(SiteSettingKey.Web).Name,
+                    Format = SiteSettingFormat.Integer
+                },
+                [SiteSettingKey.Badges.MaxFileSize] =
+                new SiteSettingDefinition
+                {
+                    Name = "Max File Size (KB)",
+                    Info = "Maximum file size of a badge in kilobytes allowed to be uploaded.",
+                    Category = typeof(SiteSettingKey.Badges).Name,
+                    Format = SiteSettingFormat.Integer
+                },
+                [SiteSettingKey.Badges.MaxDimension] =
+                new SiteSettingDefinition
+                {
+                    Name = "Max Dimension",
+                    Info = "Maximum size in pixels (width or height) allowed for a badge.",
+                    Category = typeof(SiteSettingKey.Badges).Name,
                     Format = SiteSettingFormat.Integer
                 }
             };