+++ src/GRA/SiteSettingKey.cs
@@ -2,6 +2,14 @@
 {
     public struct SiteSettingKey : System.IEquatable<SiteSettingKey>
     {
+        public struct Badges : System.IEquatable<Badges>
+        {
+            public static readonly string MaxFileSize = "Badges.MaxFileSize";
+            public static readonly string MaxDimension = "Badges.MaxDimension";
+
+            public bool Equals(Badges other) { return true; }
+        }
+
         public struct Challenges : System.IEquatable<Challenges>
         {
             public static readonly string HideUntilRegistrationOpen
@@ -19,6 +27,9 @@
 
             public static readonly string HideUntilRegistrationOpen
                 = "Events.HideUntilRegistrationOpen";
+
+            public static readonly string HideUntilProgramOpen
+                = "Events.HideUntilProgramOpen";
 
             public static readonly string RequireBadge = "Events.RequireBadge";
             public bool Equals(Events other) { return true; }